fix default features in runtime interface and make separate handling for unix in lazy-pages
